# public/doc/api/openapi.yaml
openapi: 3.1.0
info:
  title: Emodex API
  version: 0.1.0
  description: HTTP API for the Emodex Next.js application.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://{host}
    description: Production
    variables:
      host:
        default: example.com
paths:
...
  /api/groups:
    get:
      tags:
        - Groups
      summary: List groups the current user belongs to
      parameters:
        - name: X-User-Id
          in: header
          required: true
          description: Authenticated user id
          schema:
            type: integer
      responses:
        '200':
          description: Group list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupWithRole'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Groups
      summary: Create a new group
      parameters:
        - name: X-User-Id
          in: header
          required: true
          description: Authenticated user id
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate membership
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/albums:
    get:
      tags:
        - Albums
      summary: List accessible albums
      description: Returns albums owned by the user and shared albums from groups the user belongs to.
      responses:
        '200':
          description: Album list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlbumListItem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Albums
      summary: Create album
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumRequest'
      responses:
        '201':
          description: Album created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a member of the specified group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Album with same name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/albums/{id}/photo-storages:
    post:
      tags:
        - Albums
      summary: Add photo storages to album in bulk
      parameters:
        - name: id
          in: path
          required: true
          description: Album identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPhotoStoragesRequest'
      responses:
        '201':
          description: Photo storages added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddPhotoStoragesResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: photo_storage name conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/groups:
    get:
      tags:
        - Groups
      summary: List groups for the current user (Auth0 session)
      description: Returns groups the authenticated user belongs to, using Auth0 session for authentication.
      responses:
        '200':
          description: Group list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MyGroup'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: message is required
        message:
          type: string
          description: Optional message field for certain errors.
...
    Group:
      type: object
      properties:
        groupId:
          type: integer
          example: 1
        groupName:
          type: string
          example: My Group
        adminUserId:
          type: integer
          example: 10
    GroupWithRole:
      allOf:
        - $ref: '#/components/schemas/Group'
        - type: object
          properties:
            myRole:
              $ref: '#/components/schemas/MembershipRole'
...
    MembershipRole:
      type: string
      enum:
        - ADMIN
        - MEMBER
    MyGroup:
      type: object
      required:
        - groupId
        - groupName
        - myRole
      properties:
        groupId:
          type: integer
          example: 1
        groupName:
          type: string
          example: Product Launch Team
        myRole:
          $ref: '#/components/schemas/MembershipRole'
    CreateGroupRequest:
      type: object
      required:
        - groupName
      properties:
        groupName:
          type: string
          example: Product Launch Team
        memberEmails:
          type: array
          items:
            type: string
            format: email
          description: Optional list of member emails to add as MEMBERS
...
    Album:
      type: object
      required:
        - id
        - albumBasicInfo
        - photoStorageSummary
        - photoStorages
      properties:
        id:
          type: string
          example: abc123
        albumBasicInfo:
          $ref: '#/components/schemas/AlbumBasicInfo'
        photoStorageSummary:
          $ref: '#/components/schemas/PhotoStorageSummary'
        photoStorages:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/PhotoStorage'
    AlbumType:
      type: string
      enum:
        - PRIVATE
        - SHARED
    AlbumListItem:
      type: object
      required:
        - id
        - name
        - albumType
        - rootPath
        - groupId
        - groupName
        - createdTags
        - photoStorageCount
      properties:
        id:
          type: string
          example: abc123
        name:
          type: string
          example: 家族アルバム
        albumType:
          $ref: '#/components/schemas/AlbumType'
        rootPath:
          type: string
          example: 家族アルバム
        groupId:
          type:
            - integer
            - 'null'
          example: null
        groupName:
          type:
            - string
            - 'null'
          example: null
        createdTags:
          type: array
          items:
            type: string
          example:
            - 家族
            - 旅行
        photoStorageCount:
          type: integer
          minimum: 0
          example: 3
    AlbumBasicInfo:
      type: object
      required:
        - albumName
        - createdAt
        - rootPath
        - albumType
        - groupId
        - groupName
        - plannedDividend
        - createdTags
        - requiredAtAlbumCreation
      properties:
        albumName:
          type: string
          example: 思い出アルバム
        createdAt:
          type: string
          format: date-time
          example: '2026-02-13T00:00:00.000Z'
        rootPath:
          type: string
          example: 家族アルバム
        albumType:
          $ref: '#/components/schemas/AlbumType'
        groupId:
          type:
            - integer
            - 'null'
          example: null
        groupName:
          type:
            - string
            - 'null'
          example: null
        plannedDividend:
          type:
            - string
            - 'null'
          format: date-time
          example: '2026-03-31T00:00:00.000Z'
        createdTags:
          type: array
          items:
            type: string
          example:
            - 家族
            - 旅行
            - イベント
        requiredAtAlbumCreation:
          type: boolean
          example: true
    PhotoStorageSummary:
      type: object
      required:
        - totalStorages
        - totalPhotos
        - totalSizeBytes
        - lastAddedAt
      properties:
        totalStorages:
          type: integer
          minimum: 0
          example: 2
        totalPhotos:
          type: integer
          minimum: 0
          example: 12
        totalSizeBytes:
          type: integer
          format: int64
          minimum: 0
          example: 1099511
        lastAddedAt:
          type:
            - string
            - 'null'
          format: date-time
          example: '2026-02-18T10:30:00.000Z'
    PhotoStorage:
      type: object
      required:
        - id
        - name
        - storagePath
        - photoCount
        - totalSizeBytes
        - tags
        - createdAt
        - photos
      properties:
        id:
          type: string
          example: ps_123
        name:
          type: string
          example: 2026-02-旅行
        storagePath:
          type: string
          example: 家族アルバム/2026-02-旅行
        photoCount:
          type: integer
          minimum: 1
          example: 8
        totalSizeBytes:
          type: integer
          format: int64
          minimum: 1
          example: 8123456
        tags:
          type: array
          items:
            type: string
          example:
            - 旅行
            - 家族
        createdAt:
          type: string
          format: date-time
          example: '2026-02-18T10:30:00.000Z'
        photos:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/PhotoStoragePhoto'
    PhotoStoragePhoto:
      type: object
      required:
        - id
        - fileName
        - blobPath
        - blobUrl
        - sizeBytes
      properties:
        id:
          type: string
          example: psp_123
        fileName:
          type: string
          example: IMG_0001.JPG
        blobPath:
          type: string
          example: 家族アルバム/2026-02-旅行/1700000000-1-IMG_0001.JPG
        blobUrl:
          type: string
          format: uri
          example: https://example.public.blob.vercel-storage.com/xxx.jpg
        contentType:
          type:
            - string
            - 'null'
          example: image/jpeg
        sizeBytes:
          type: integer
          format: int64
          minimum: 1
          example: 2048000
    CreateAlbumRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          example: 家族アルバム
        albumType:
          $ref: '#/components/schemas/AlbumType'
        groupId:
          type: integer
          description: Required when albumType is SHARED
          example: 1
        plannedDividend:
          type: string
          format: date-time
          example: '2026-03-31T00:00:00.000Z'
        createdTags:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 50
          example:
            - 家族
            - 旅行
        requiredAtAlbumCreation:
          type: boolean
          example: false
    AddPhotoStoragesRequest:
      type: object
      required:
        - name
        - files
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          example: 2026-02-旅行
        tags:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 50
          example:
            - 旅行
            - 家族
        files:
          type: array
          minItems: 1
          maxItems: 200
          items:
            type: object
            required:
              - fileName
              - blobPath
              - blobUrl
              - sizeBytes
            properties:
              fileName:
                type: string
                example: IMG_0001.JPG
              blobPath:
                type: string
                example: 家族アルバム/2026-02-旅行/1700000000-1-IMG_0001.JPG
              blobUrl:
                type: string
                format: uri
                example: https://example.public.blob.vercel-storage.com/xxx.jpg
              contentType:
                type: string
                example: image/jpeg
              sizeBytes:
                type: integer
                minimum: 1
                example: 2048000
    AddPhotoStoragesResponse:
      type: object
      required:
        - albumId
        - photoStorage
      properties:
        albumId:
          type: string
          example: abc123
        photoStorage:
          $ref: '#/components/schemas/PhotoStorage'
...